Locale.txt improvements (#82)
* Create locale.txt when generating for console

* Logic fixes

* Locale.txt improvements
A locale.txt file will now only be generated if the selected game region is different to the console region.
If a locale.txt file does need to be generated a new option to select the language set by the file will appear. If available, the default value will be the system's language setting, otherwise it will default to English.